+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Upcoming
 
-- handle dynamic styling (cva, clsx, tw, nunjucks, react/vue/etc. conditional syntax)
+- handle dynamic styling 
 
+
+### 0.0.8
+
+- Support for additional prefixes ex. `twMerge(`, `cva(` and `clsx(`
+- Add missing classes: ring, ring-inset, ring-offset to borders category
 
 ### 0.0.72